requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
PySocks==1.7.1
python-dateutil==2.9.0.post0
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
